Scaled textures for cuboids
Set Cuboid (and plane) textures to scale such that all cuboid objects with the same texture will look the same such that the texture is never stretched and appears the same size on any of those objects. Does not work too well with curved surfaces so currently only applied to cuboids and planes.